requirements.txt
python-dotenv==0.15.0
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt